pyproject.toml
[project]
name = "pykit_vortexdude"
version = "0.0.1"
authors = [
  { name="Nitin Namdev", email="itsmyidbro@gmail.com" },
]
description = "A Example Utils package"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "PyYAML",
    "requests"
]


[project.urls]
Homepage = "https://github.com/vortexdude/PyKit"
Issues = "https://github.com/vortexdude/PyKit/issues"


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
